+++ src/apps/jones-dao/contracts/index.ts
@@ -4,10 +4,13 @@ import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { ContractFactory } from '~contract/contracts';
 import { Network } from '~types/network.interface';
 
-import { JonesMillinerV2__factory } from './ethers';
-import { JonesStakingRewards__factory } from './ethers';
-import { JonesStakingRewardsFactory__factory } from './ethers';
-import { JonesVault__factory } from './ethers';
+import {
+  JonesMetavault__factory,
+  JonesMillinerV2__factory,
+  JonesStakingRewards__factory,
+  JonesStakingRewardsFactory__factory,
+  JonesVault__factory,
+} from './ethers';
 
 // eslint-disable-next-line
 type ContractOpts = { address: string; network: Network };
@@ -18,6 +21,9 @@ export class JonesDaoContractFactory extends ContractFactory {
     super((network: Network) => appToolkit.getNetworkProvider(network));
   }
 
+  jonesMetavault({ address, network }: ContractOpts) {
+    return JonesMetavault__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  }
   jonesMillinerV2({ address, network }: ContractOpts) {
     return JonesMillinerV2__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
@@ -32,6 +38,7 @@ export class JonesDaoContractFactory extends ContractFactory {
   }
 }
 
+export type { JonesMetavault } from './ethers';
 export type { JonesMillinerV2 } from './ethers';
 export type { JonesStakingRewards } from './ethers';
 export type { JonesStakingRewardsFactory } from './ethers';