+++ src/apps/lemma-finance/optimism/lemma-finance.synth.token-fetcher.ts
@@ -1,120 +1,102 @@
 import { Inject } from '@nestjs/common';
-import _ from 'lodash';
 
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
-import { Register } from '~app-toolkit/decorators';
-import { buildDollarDisplayItem } from '~app-toolkit/helpers/presentation/display-item.present';
-import { getAppAssetImage } from '~app-toolkit/helpers/presentation/image.present';
-import { ContractType } from '~position/contract.interface';
-import { PositionFetcher } from '~position/position-fetcher.interface';
-import { AppTokenPosition } from '~position/position.interface';
-import { Network } from '~types/network.interface';
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+import { AppTokenTemplatePositionFetcher } from '~position/template/app-token.template.position-fetcher';
+import {
+  GetAddressesParams,
+  GetUnderlyingTokensParams,
+  DefaultAppTokenDataProps,
+  GetPricePerShareParams,
+} from '~position/template/app-token.template.types';
 
-import { LemmaFinanceContractFactory } from '../contracts';
-import { LEMMA_FINANCE_DEFINITION } from '../lemma-finance.definition';
+import { LemmaFinanceContractFactory, LemmaSynth } from '../contracts';
 
-const synthAddresses = [
+export type LemmaFinanceSynthDefinition = {
+  address: string;
+  perpAddress: string;
+};
+
+const synthDefinitions = [
   {
-    synthAddress: '0x3bc414fa971189783acee4dee281067c322e3412',
+    address: '0x3bc414fa971189783acee4dee281067c322e3412', // lETH
     perpAddress: '0x29b159ae784accfa7fb9c7ba1de272bad75f5674',
   },
   {
-    synthAddress: '0x8a641696caf0f59bb7a53cf8d2dc943ed95229a6',
+    address: '0x8a641696caf0f59bb7a53cf8d2dc943ed95229a6', // lWBTC
     perpAddress: '0xe161c6c9f2fc74ac97300e6f00648284d83cbd19',
   },
   {
-    synthAddress: '0x5c39a4a368ab3c3239d20eb4219e0361bd2ad092',
+    address: '0x5c39a4a368ab3c3239d20eb4219e0361bd2ad092', // lLINK
     perpAddress: '0xdd4d71d3563c24e38525661896e1d01fd8c2c9a5',
   },
   {
-    synthAddress: '0x546ba811099883bef35fa360e7ded8af439831f3',
+    address: '0x546ba811099883bef35fa360e7ded8af439831f3', // lCRV
     perpAddress: '0x119f85ecfcfbc1d7033d266192626202df7dbdf2',
   },
   {
-    synthAddress: '0xd1a988b024c55d7baabb07fd531d63a4e19e3b4c',
+    address: '0xd1a988b024c55d7baabb07fd531d63a4e19e3b4c', // lPERP
     perpAddress: '0x13c214b430fe304c4c6437f3564a690cd4e4f23b',
   },
   {
-    synthAddress: '0xa7c657a94eb9571f4e94f49943af1130e6d7337c',
+    address: '0xa7c657a94eb9571f4e94f49943af1130e6d7337c', // lAAVE
     perpAddress: '0xfe1eb36d31ead771fd5e051ee8cc424db6416567',
   },
 ];
 
-const appId = LEMMA_FINANCE_DEFINITION.id;
-const groupId = LEMMA_FINANCE_DEFINITION.groups.synth.id;
-const network = Network.OPTIMISM_MAINNET;
+@PositionTemplate()
+export class OptimismLemmaFinanceSynthTokenFetcher extends AppTokenTemplatePositionFetcher<
+  LemmaSynth,
+  DefaultAppTokenDataProps,
+  LemmaFinanceSynthDefinition
+> {
+  groupLabel = 'Synths';
 
-@Register.TokenPositionFetcher({ appId, groupId, network })
-export class OptimismLemmaFinanceLemmaSynthTokenFetcher implements PositionFetcher<AppTokenPosition> {
   constructor(
-    @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
-    @Inject(LemmaFinanceContractFactory) private readonly contractFactory: LemmaFinanceContractFactory,
-  ) {}
+    @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
+    @Inject(LemmaFinanceContractFactory) protected readonly contractFactory: LemmaFinanceContractFactory,
+  ) {
+    super(appToolkit);
+  }
 
-  async getPositions() {
-    const baseTokens = await this.appToolkit.getBaseTokenPrices(network);
-    const multicall = this.appToolkit.getMulticall(network);
+  getContract(address: string): LemmaSynth {
+    return this.contractFactory.lemmaSynth({ address, network: this.network });
+  }
 
-    const tokens = await Promise.all(
-      synthAddresses.map(async position => {
-        const contract = this.contractFactory.synth({
-          address: position.synthAddress,
-          network,
-        });
+  async getDefinitions() {
+    return synthDefinitions;
+  }
 
-        const perpContract = this.contractFactory.perpLemma({
-          address: position.perpAddress,
-          network,
-        });
+  async getAddresses({ definitions }: GetAddressesParams<LemmaFinanceSynthDefinition>) {
+    return definitions.map(v => v.address);
+  }
 
-        const [name, symbol, decimals, supplyRaw, collateral, usdc, indexPrice] = await Promise.all([
-          multicall.wrap(contract).name(),
-          multicall.wrap(contract).symbol(),
-          multicall.wrap(contract).decimals(),
-          multicall.wrap(contract).totalSupply(),
-          multicall.wrap(contract).tailCollateral(),
-          multicall.wrap(perpContract).usdc(),
-          multicall.wrap(perpContract).getIndexPrice(),
-        ]);
+  async getUnderlyingTokenDefinitions({
+    definition,
+    multicall,
+  }: GetUnderlyingTokensParams<LemmaSynth, LemmaFinanceSynthDefinition>) {
+    const synthContract = this.contractFactory.lemmaSynth({ address: definition.address, network: this.network });
+    const perpContract = this.contractFactory.lemmaPerp({ address: definition.perpAddress, network: this.network });
 
-        const supply = Number(supplyRaw) / 10 ** decimals;
-        const usdcToken = baseTokens.find(x => x.address.toLowerCase() === usdc.toLowerCase());
-        const collateralToken = baseTokens.find(x => x.address.toLowerCase() === collateral.toLowerCase());
-        if (!usdcToken || !collateralToken) return null;
+    const [collateralTokenAddress, usdcAddress] = await Promise.all([
+      multicall.wrap(synthContract).tailCollateral(),
+      multicall.wrap(perpContract).usdc(),
+    ]);
 
-        const tokens = [usdcToken, collateralToken];
-        const price = Number(indexPrice) / 10 ** decimals;
-        const pricePerShare = 1;
+    return [
+      { address: usdcAddress, network: this.network },
+      { address: collateralTokenAddress, network: this.network },
+    ];
+  }
 
-        const label = `${name} (${symbol})`;
-        const images = [getAppAssetImage(appId, symbol)];
-        const secondaryLabel = buildDollarDisplayItem(price);
-
-        const token: AppTokenPosition = {
-          type: ContractType.APP_TOKEN,
-          appId,
-          groupId,
-          address: position.synthAddress,
-          network,
-          symbol,
-          decimals,
-          supply,
-          tokens,
-          price,
-          pricePerShare,
-          dataProps: {
-            liquidity: supply * price,
-          },
-          displayProps: {
-            label,
-            images,
-            secondaryLabel,
-          },
-        };
-        return token;
-      }),
-    );
-
-    return _.compact(tokens);
+  async getPricePerShare({
+    definition,
+    multicall,
+    appToken,
+  }: GetPricePerShareParams<LemmaSynth, DefaultAppTokenDataProps, LemmaFinanceSynthDefinition>) {
+    const perpContract = this.contractFactory.lemmaPerp({ address: definition.perpAddress, network: this.network });
+    const indexPriceRaw = await multicall.wrap(perpContract).getIndexPrice();
+    const indexPrice = Number(indexPriceRaw) / 10 ** appToken.decimals;
+    return [indexPrice, 0];
   }
 }