+++ src/apps/mycelium/arbitrum/mycelium.es-myc.token-fetcher.ts
@@ -1,24 +1,33 @@
 import { Inject } from '@nestjs/common';
+import 'moment-duration-format';
 
-import { Register } from '~app-toolkit/decorators';
-import { PositionFetcher } from '~position/position-fetcher.interface';
-import { AppTokenPosition } from '~position/position.interface';
-import { Network } from '~types/network.interface';
+import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+import { Erc20 } from '~contract/contracts';
+import { AppTokenTemplatePositionFetcher } from '~position/template/app-token.template.position-fetcher';
 
-import { MyceliumEsMycTokenHelper } from '../helpers/mycelium.es-myc.token-helper';
-import { MYCELIUM_DEFINITION } from '../mycelium.definition';
+import { MyceliumContractFactory } from '../contracts';
 
-const appId = MYCELIUM_DEFINITION.id;
-const groupId = MYCELIUM_DEFINITION.groups.esMyc.id;
-const network = Network.ARBITRUM_MAINNET;
+@PositionTemplate()
+export class ArbitrumMyceliumEsMycTokenFetcher extends AppTokenTemplatePositionFetcher<Erc20> {
+  groupLabel = 'esMYC';
 
-@Register.TokenPositionFetcher({ appId, groupId, network })
-export class ArbitrumMyceliumEsMycTokenFetcher implements PositionFetcher<AppTokenPosition> {
-  constructor(@Inject(MyceliumEsMycTokenHelper) private readonly myceliumEsMycTokenHelper: MyceliumEsMycTokenHelper) {}
+  constructor(
+    @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
+    @Inject(MyceliumContractFactory) protected readonly contractFactory: MyceliumContractFactory,
+  ) {
+    super(appToolkit);
+  }
 
-  async getPositions() {
-    return this.myceliumEsMycTokenHelper.getTokens({
-      network,
-    });
+  getContract(address: string): Erc20 {
+    return this.contractFactory.erc20({ network: this.network, address });
+  }
+
+  async getAddresses() {
+    return ['0x7cec785fba5ee648b48fbffc378d74c8671bb3cb'];
+  }
+
+  async getUnderlyingTokenDefinitions() {
+    return [{ address: '0xc74fe4c715510ec2f8c61d70d397b32043f55abe', network: this.network }];
   }
 }